changes made as per the PR review
