+++ src/@/sidebar/sidebar.model.ts
@@ -1,4 +1,4 @@
-import { combine, createEvent, createStore, restore, sample } from 'effector';
+import { combine, createEvent, createStore, merge, restore, sample } from 'effector';
 import { ConnectivityDistributionNames, getConnectivityLogicalValues, LayerDistributionUnit } from './ui/global-and-country-view-components/container/layer-view.constant';
 
 import { $admin1Code, $country, $countryActiveLayersDataById, $countryBenchmark, $countryCode, $countryConnectivityNames, $countryIdToCode, $countrySearchString } from '~/@/country/country.model';
@@ -512,357 +512,3 @@ $isLoadingTimeplayer.reset($isTimeplayer)
 $sidebarHeight.reset([router.historyUpdated, $showLegend])
 
 $showAdvancedFilter.reset([$countryCode, $admin1Code, $countrySearchString])
-
-type Layer = { id?: string | number; code?: string; name?: string } | null;
-
-/**
- * Guards for URL handling:
- * - `initialized` prevents applying URL params more than once.
- * - `urlWriteEnabled` prevents URL writes during initial hydration/apply.
- */
-let initialized = false;
-let urlWriteEnabled = false;
-
-/** Sanitize a string for use in URL tokens (lowercase, alnum, dash, underscore). */
-function sanitize(str: string): string {
-  return str
-    .trim()
-    .toLowerCase()
-    .replace(/[^a-z0-9_-]+/g, '-') // allow alnum, dash, underscore
-    .replace(/-+/g, '-')
-    .replace(/^-|-$/g, '');
-}
-
-/**
- * FLAG_META maps logical flags to:
- * - a fieldName used in the combined $currentLayer output,
- * - the urlParam and optional legendUrlParam names,
- * - functions to get the actual layer and derive legends,
- * - a legendSetter to apply legend tokens (Set<string>) to your stores/events,
- * - clearOnAbsent controls whether to clear legend state when param is absent (default true).
- *
- * Keep legendSetter implementations aligned with your actual effector events/stores.
- */
-const FLAG_META: Record<
-  string,
-  {
-    fieldName: string;
-    urlParam: string;
-    legendUrlParam?: string;
-    getLayer: (foundLayer: Layer) => Layer;
-    getLegends?: (opts: {
-      staticLegendsSelected: string[];
-      coverageStatusAll: Record<string, boolean>;
-      liveLayerLegendsStatus: Record<string, boolean>;
-    }) => string[];
-    legendSetter?: (tokens: Set<string>) => void;
-    clearOnAbsent?: boolean;
-  }
-> = {
-  isLive: {
-    fieldName: 'liveLayer',
-    urlParam: 'layer__live',
-    legendUrlParam: 'layer__include_legend__connectivity_speed',
-    getLayer: (f) => f,
-    getLegends: ({ liveLayerLegendsStatus }) =>
-      Object.keys(liveLayerLegendsStatus || {}).filter((k) => !!(liveLayerLegendsStatus as any)[k]),
-    legendSetter: (set) => {
-      changeConnectivitySpeedGood(Boolean(set.has('good')));
-      changeConnectivitySpeedModerate(Boolean(set.has('moderate')));
-      changeConnectivitySpeednoInternet(Boolean(set.has('no_internet') || set.has('no-internet') || set.has('bad')));
-      changeConnectivitySpeedUnknown(Boolean(set.has('unknown')));
-    },
-    clearOnAbsent: false,
-  },
-
-  isStatic: {
-    fieldName: 'staticLayer',
-    urlParam: 'layer__static',
-    legendUrlParam: 'layer__include_legend__coverage_status',
-    getLayer: (f) => f,
-    getLegends: ({ coverageStatusAll }) =>
-      Object.keys(coverageStatusAll || {}).filter((k) => !!(coverageStatusAll as any)[k]),
-    legendSetter: (set) => {
-      changeCoverage5g4g(Boolean(set.has('good')));
-      changeCoverage3g2g(Boolean(set.has('moderate')));
-      changeCoverageNoCoverage(Boolean(set.has('bad')));
-      changeCoverageUnknown(Boolean(set.has('unknown')));
-    },
-    clearOnAbsent: false,
-  },
-
-  isSchoolStatus: {
-    fieldName: 'schoolStatusLayer',
-    urlParam: 'layer__schoolStatus',
-    legendUrlParam: 'layer__include_legend__school_status',
-    getLayer: () => (typeof SCHOOL_STATUS_LAYER !== 'undefined' ? (SCHOOL_STATUS_LAYER as Layer) : null),
-    getLegends: ({ staticLegendsSelected }) => (Array.isArray(staticLegendsSelected) ? staticLegendsSelected.slice() : []),
-    legendSetter: (set) => {
-      staticLegendsSelection(Array.from(set));
-    },
-    clearOnAbsent: false,
-  },
-};
-
-/**
- * Combine $currentLayer with all legend-related stores.
- * Add legend-related stores here so $currentLayer contains both selected layer and current legend state.
- */
-export const $currentLayer = combine(
-  {
-    layers: $layersList,
-    selectedId: $selectedLayerId,
-    typeUtils: $currentLayerTypeUtils,
-    // legend sources:
-    staticLegendsSelected: $staticLegendsSelected,
-    coverageStatusAll: $coverageStatusAll,
-    liveLayerLegendsStatus: $liveLayerLegendsStatus,
-    activeLayerByCountryCode: $activeLayerByCountryCode,
-    currentDefaultLayerIdForUI: $currentDefaultLayerIdForUI,
-  },
-  (payload): Record<string, any> => {
-    const {
-      layers,
-      selectedId,
-      typeUtils,
-      staticLegendsSelected,
-      coverageStatusAll,
-      liveLayerLegendsStatus,
-      activeLayerByCountryCode,
-      currentDefaultLayerIdForUI,
-    } = payload as {
-      layers: any[];
-      selectedId: string | number | null;
-      typeUtils: any;
-      staticLegendsSelected: string[];
-      coverageStatusAll: Record<string, boolean>;
-        liveLayerLegendsStatus: Record<string, boolean>;
-      activeLayerByCountryCode: Record<string, boolean>;
-        currentDefaultLayerIdForUI: number | null;
-    };
-
-    if (!typeUtils) return {};
-
-    // find the selected layer object from the list
-    const findLayer = (): Layer => {
-      if (!selectedId || !Array.isArray(layers)) return null;
-      return layers.find((l: any) => `${l?.id}` === `${selectedId}`) ?? null;
-    };
-    const found = findLayer();
-
-    const result: Record<string, any> = {};
-
-    // determine if any flag is explicitly enabled; use schoolStatus as default when none enabled
-    const anyFlagEnabled = Object.keys(FLAG_META).some((flag) => Boolean((typeUtils as any)[flag]));
-
-    Object.keys(FLAG_META).forEach((flag) => {
-      const isEnabled = Boolean((typeUtils as any)[flag]) || (!anyFlagEnabled && flag === 'isSchoolStatus');
-      if (!isEnabled) return;
-
-      const meta = FLAG_META[flag];
-
-      // set layer object for this flag
-      result[meta.fieldName] = meta.getLayer(found);
-
-      // compute and attach legend tokens for URL writing if provider exists
-      if (meta.getLegends) {
-        const legends = meta.getLegends({
-          staticLegendsSelected,
-          coverageStatusAll,
-          liveLayerLegendsStatus,
-        }) || [];
-        result[`${meta.fieldName}Legends`] = Array.isArray(legends) ? legends.map(String) : [];
-      }
-    });
-
-    return result;
-  }
-);
-
-/**
- * Event that accepts an object shaped like $currentLayer's output and writes URL params.
- * The watcher will build and replace URL using history.replaceState.
- */
-const updateUrlWithLayers = createEvent<Record<string, any>>();
-
-updateUrlWithLayers.watch((layersObj) => {
-  if (typeof window === 'undefined') return;
-  if (!urlWriteEnabled) return;
-
-  try {
-    const url = new URL(window.location.href);
-    const params = url.searchParams;
-
-    // Write layer__ params
-    Object.values(FLAG_META).forEach((meta) => {
-      const layer = (layersObj as any)[meta.fieldName] as Layer | undefined;
-      const key = meta.urlParam?.toLowerCase();
-      if (layer && layer.id != null && key) {
-        const rawCode =
-          (typeof layer.code === 'string' && layer.code.trim() !== '')
-            ? layer.code
-            : (typeof layer.name === 'string' && layer.name.trim() !== '')
-              ? layer.name
-              : String(layer.id);
-        params.set(key, `${String(layer.id)}_${sanitize(String(rawCode))}`);
-      } else if (key) {
-        params.delete(key);
-      }
-    });
-
-    // Write legend params (if present)
-    Object.values(FLAG_META).forEach((meta) => {
-      const legendKey = meta.legendUrlParam?.toLowerCase();
-      if (!legendKey) return;
-      const legends = (layersObj as any)[`${meta.fieldName}Legends`] as string[] | undefined;
-      if (Array.isArray(legends) && legends.length > 0) {
-        const sanitized = legends
-          .map((l) => String(l).trim())
-          .filter(Boolean)
-          .map((l) => sanitize(l))
-          .filter(Boolean);
-        if (sanitized.length) params.set(legendKey, sanitized.join(','));
-        else params.delete(legendKey);
-      } else {
-        params.delete(legendKey);
-      }
-    });
-
-    // global 'layer' flag exists if any layer__ params present
-    const hasAnyLayerParam = Array.from(params.keys()).some((k) => k.toLowerCase().startsWith('layer__'));
-    if (hasAnyLayerParam) params.set('layer', 'true');
-    else params.delete('layer');
-
-    const newUrl = url.pathname + (params.toString() ? `?${params.toString()}` : '') + url.hash;
-    const currentUrl = window.location.pathname + window.location.search + window.location.hash;
-    if (newUrl !== currentUrl) window.history.replaceState(null, '', newUrl);
-  } catch (err) {
-    // swallow errors to avoid breaking the app flow
-  }
-});
-
-export const triggerUpdateUrl = createEvent<void>();
-
-// Sample to trigger URL write when required. Keep onSelectMainLayer in the clock array if you want writes on main-layer selection.
-sample({
-  source: $currentLayer,
-  clock: [triggerUpdateUrl, onSelectMainLayer],
-  target: updateUrlWithLayers,
-});
-
-/* --------------------- URL parsing & apply (refactored) --------------------- */
-
-/* Parse "<id>_<raw>" into parts. If underscore missing, raw is empty string. */
-function parseLayerParamValue(value: string | null) {
-  if (!value) return null;
-  const idx = value.indexOf('_');
-  if (idx === -1) return { id: value, raw: '' };
-  return { id: value.slice(0, idx), raw: value.slice(idx + 1) };
-}
-
-/* Convert "a,b,c" -> ['a','b','c'] (lowercased, trimmed). */
-function parseLegendParamValue(value: string | null) {
-  if (!value) return [];
-  return value
-    .split(',')
-    .map((s) => String(s).trim())
-    .filter(Boolean)
-    .map((s) => s.toLowerCase());
-}
-
-/* Read layer-related params from current URL when the global 'layer' flag is 'true'. */
-function readLayerParams(): Record<string, string> {
-  if (typeof window === 'undefined') return {};
-  const url = new URL(window.location.href);
-  const params = url.searchParams;
-  const layerFlag = params.get('layer');
-  if (layerFlag !== 'true') return {}; // no layer intent
-  const out: Record<string, string> = {};
-  params.forEach((value, key) => {
-    const k = key.toLowerCase();
-    if (k.startsWith('layer')) out[k] = value;
-  });
-  return out;
-}
-
-/* Convert raw legend param string -> Set of normalized tokens (underscores instead of dashes). */
-const tokensToSet = (raw?: string) => {
-  if (!raw) return new Set<string>();
-  return new Set(
-    parseLegendParamValue(raw)
-      .map((t) => t.replace(/-/g, '_'))
-      .map((t) => t.trim())
-      .filter(Boolean)
-  );
-};
-
-/**
- * Apply URL params to app state once during initialization:
- * - reads known layer and legend params,
- * - dispatches existing effector events to set initial selection/legend state,
- * - enables URL writer after initial hydration is complete.
- *
- * Call this once on app init (client-side) when effector stores/events are ready.
- */
-export const applyUrlParams = () => {
-  if (initialized) return;
-  initialized = true;
-
-  const layerParams = readLayerParams();
-  if (!layerParams || Object.keys(layerParams).length === 0) {
-    // no params to apply — allow URL writer to run for future changes
-    urlWriteEnabled = true;
-    return;
-  }
-
-  /* helper to read known params in a case-insensitive way */
-  const getParam = (k: string) => layerParams[k.toLowerCase()] ?? null;
-
-  const schoolRaw = getParam('layer__schoolstatus');
-  const staticRaw = getParam('layer__static');
-  const liveRaw = getParam('layer__live');
-
-  const schoolId = schoolRaw ? Number(parseLayerParamValue(schoolRaw)?.id) : null;
-  const staticId = staticRaw ? Number(parseLayerParamValue(staticRaw)?.id) : null;
-  const liveId = liveRaw ? Number(parseLayerParamValue(liveRaw)?.id) : null;
-
-  // Prefer live -> static for setting the preferred layer id for UI
-  setUrlPreferredLayer({
-    hasIdInUrl: String(layerParams['layer']) === 'true',
-    layerId: liveId ?? staticId,
-  });
-
-  // apply school status selection (send null if invalid)
-  onSelectSchoolStatusLayer(typeof schoolId === 'number' && !Number.isNaN(schoolId) && schoolId !== 0 ? schoolId : null);
-
-  // decide preferred layer id and dispatch checks/selection
-  const urlPreferredLayer = (typeof liveId === 'number' && !Number.isNaN(liveId) && liveId !== 0)
-    ? liveId
-    : (typeof staticId === 'number' && !Number.isNaN(staticId) && staticId !== 0)
-      ? staticId
-      : null;
-
-  if (urlPreferredLayer) {
-    checkConnectivityBenchmark(urlPreferredLayer);
-    onSelectMainLayer(urlPreferredLayer);
-  }
-
-  // apply legend setters (call with Set of tokens) or clear when absent (unless clearOnAbsent is false)
-  Object.values(FLAG_META).forEach((meta) => {
-    const key = (meta.legendUrlParam || '').toLowerCase();
-    if (!key) return;
-
-    const raw = layerParams[key] ?? null;
-    if (raw) {
-      const set = tokensToSet(raw as string);
-      if (meta.legendSetter) meta.legendSetter(set);
-    } else {
-      if (meta.clearOnAbsent !== false && meta.legendSetter) {
-        meta.legendSetter(new Set());
-      }
-    }
-  });
-
-  // enable URL writer now that initial state is applied
-  urlWriteEnabled = true;
-};
